requirements.txt
transformers
torch
pyyaml
